DOS wip

+++ RayCaster/RayCasterFloat.h
@@ -5,8 +5,8 @@
 class RayCasterFloat : public RayCaster
 {
 public:
-    void Start(uint16_t playerX, uint16_t playerY, int16_t playerA);
-    void Trace(uint16_t screenX, uint8_t* screenY, uint8_t* textureNo, uint8_t* textureX, uint16_t* textureY, uint16_t* textureStep);
+    void Start(uint32_t playerX, uint32_t playerY, int32_t playerA);
+    void Trace(uint32_t screenX, uint8_t* screenY, uint8_t* textureNo, uint8_t* textureX, uint32_t* textureY, uint32_t* textureStep);
 
     RayCasterFloat();
     ~RayCasterFloat();
